+++ include/argv.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include "mruby.h"
+
+void populate_argv(mrb_state*, int, char**);